// web/src/shared/api/generated/model/multiUnitExamUnitOptionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface MultiUnitExamUnitOptionResult {
  unitId: string;
  name: string;
  hasBlueprint: boolean;
  isSubjectDefault: boolean;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  servableCount: number | string;
}